requirements.txt
shotgun-api3 @ git+https://github.com/shotgunsoftware/python-api.git@v3.6.1
pytest~=8.3.2
python-dotenv~=1.0.1